Cleanup tear-down methods to stabilize test cases
The refresh call was added as part of the JUnit4 migration in an attempt
to stabilize some of the failing test cases. Instead, this seems to
have destabilized previously stable tests. In order to stabilize
ProjectClassLoaderTest, we now call FileUtils.forceDelete() instead of
FileUtils.deleteQuietly().
We also need to ensure that resources are ALWAYS deleted once a test
finishes, even if the test itself has failed. Otherwise those files may
leak into the next test.