Refactor selection handling in GraphicalRobot
Similar to other GEF replay tools it is not possible to select edit
parts that are outside the visible area. But this functionality is
required as part of the GEF migration.

To solve this problem, the absolute coordinates of the robot need to be
converted to relative coordinates, as well as scrolling the viewer so
that the selection becomes visible.